exercise 4 completed

+++ 4/backend/controllers/blogs.js
@@ -1,10 +1,13 @@
 const blogsRouter = require('express').Router();
+
 const Blog = require('../models/blog');
 
 // API Endpoints
 blogsRouter.get('/', async (request, response) => {
 	try {
-		const blogs = await Blog.find({});
+		const blogs = await Blog.find({})
+			.find({})
+			.populate('user', { username: 1, name: 1 });
 		response.json(blogs);
 	} catch (error) {
 		// Handle any errors that occur during the database query
@@ -32,8 +35,16 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
 	try {
-		const blog = new Blog(request.body);
+		const user = request.user;
+
+		if (!user) {
+			return response.status(401).json({ error: 'invalid or missing token' });
+		}
+
+		const blog = new Blog({ ...request.body, user: user.id });
 		const result = await blog.save();
+		user.blogs = user.blogs.concat(result._id);
+		await user.save();
 		response.status(201).json(result);
 	} catch (error) {
 		// Handle any errors that occur during the save operation
@@ -49,8 +60,21 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
 	try {
-		await Blog.findByIdAndDelete(request.params.id);
-		response.status(204).end();
+		const user = request.user;
+
+		if (!user) {
+			return response.status(401).json({ error: 'invalid or missing token' });
+		}
+
+		const blog = await Blog.findById(request.params.id);
+		if (blog.user.toString() === request.user.id) {
+			await Blog.findByIdAndDelete(request.params.id);
+			response.status(204).end();
+		} else {
+			return response
+				.status(401)
+				.json({ error: 'Unauthorized to delete the blog' });
+		}
 	} catch (error) {
 		console.error('Error deleting blog:', error);
 		response.status(500).json({ error: 'Internal Server Error' });
